- Moved utility functions to a separate `utils.js` file for better modularity. - Updated the caching mechanism to extend the cache TTL from 5 to 10 minutes for improved performance. - Refactored request handling methods to utilize the new utility functions, enhancing readability and maintainability.
