+++ src/index.ts
@@ -30,6 +30,14 @@ import { z } from 'zod';
 import * as fs from 'fs';
 import * as path from 'path';
 import * as os from 'os';
+import {
+  generateRequestId,
+  sanitizeFilename,
+  toBooleanSafe,
+  generateExportMessage,
+  performHybridSearch,
+  performExactSearch,
+} from './utils.js';
 
 // Custom error enum
 enum ErrorCode {
@@ -102,7 +110,7 @@ class MetabaseServer {
   };
   // Add caching for expensive operations
   private cardsCache: { data: any[] | null; timestamp: number | null } = { data: null, timestamp: null };
-  private readonly CACHE_TTL_MS = 5 * 60 * 1000; // 5 minutes cache
+  private readonly CACHE_TTL_MS = 10 * 60 * 1000; // 10 minutes cache
   private readonly REQUEST_TIMEOUT_MS = 30000; // 30 seconds timeout
 
   constructor() {
@@ -352,7 +360,7 @@ class MetabaseServer {
    */
   private setupResourceHandlers() {
     this.server.setRequestHandler(ListResourcesRequestSchema, async (_request) => {
-      this.logInfo('Processing request to list resources', { requestId: this.generateRequestId() });
+      this.logInfo('Processing request to list resources', { requestId: generateRequestId() });
       await this.getSessionToken();
 
       try {
@@ -410,7 +418,7 @@ class MetabaseServer {
 
     // Read resource
     this.server.setRequestHandler(ReadResourceRequestSchema, async (request) => {
-      const requestId = this.generateRequestId();
+      const requestId = generateRequestId();
       this.logInfo('Processing request to read resource', {
         requestId,
         uri: request.params?.uri
@@ -499,100 +507,6 @@ class MetabaseServer {
         );
       }
     });
-  }
-
-  /**
-   * Generate a unique request ID
-   */
-  private generateRequestId(): string {
-    return Math.random().toString(36).substring(2, 15);
-  }
-
-  /**
-   * Sanitize filename to prevent path traversal attacks
-   */
-  private sanitizeFilename(filename: string | undefined): string {
-    if (!filename || typeof filename !== 'string') {
-      return '';
-    }
-
-    // Remove path separators and other dangerous characters
-    // Keep only alphanumeric, hyphens, underscores, spaces, and dots
-    return filename
-      .replace(/[/\\:*?"<>|]/g, '') // Remove path separators and invalid filename chars
-      .replace(/\.\./g, '') // Remove parent directory references
-      .replace(/^\.+/, '') // Remove leading dots
-      .trim()
-      .substring(0, 255); // Limit length to prevent filesystem issues
-  }
-
-  /**
-   * Convert parameter to boolean safely
-   */
-  private toBooleanSafe(value: unknown): boolean {
-    if (typeof value === 'boolean') {
-      return value;
-    }
-    if (typeof value === 'string') {
-      return value.toLowerCase() === 'true';
-    }
-    if (typeof value === 'number') {
-      return value !== 0;
-    }
-    return Boolean(value);
-  }
-
-  /**
-   * Generate standardized export result message
-   */
-  private generateExportMessage(
-    format: string,
-    query: string,
-    databaseId: number,
-    rowCount: number,
-    fileSize: string,
-    saveFile: boolean,
-    savedFilePath: string,
-    filename: string,
-    fileSaveError?: string
-  ): string {
-    const queryPreview = query.length > 100 ? `${query.substring(0, 100)}...` : query;
-
-    let statusMessage = '';
-    if (saveFile) {
-      if (fileSaveError) {
-        statusMessage = `\nFile Save Status: FAILED - ${fileSaveError}\nFallback: Use manual copy-paste method below\n`;
-      } else {
-        statusMessage = `\nFile Save Status: SUCCESS\nFile Location: ${savedFilePath}\nDownloads Folder: Available for use\n`;
-      }
-    }
-
-    const formatUpper = format.toUpperCase();
-
-    return `# Query Export Results (${formatUpper} Format)
-
-Query: ${queryPreview}
-Database ID: ${databaseId}
-${format === 'xlsx' ? `File Size: ${fileSize} bytes` : `Rows Exported: ${rowCount.toLocaleString()}`}
-Export Method: Metabase high-capacity API (supports up to 1M rows)${statusMessage}
-
-## Manual Save Instructions${saveFile && !fileSaveError ? ' (Alternative Method)' : ''}:
-
-1. Select all the ${formatUpper} content below${format === 'csv' ? ' (between the ```csv markers)' : ''}
-2. Copy the selected text (Cmd+C / Ctrl+C)
-3. Open a ${format === 'xlsx' ? 'spreadsheet application' : format === 'json' ? 'text editor' : 'text editor or spreadsheet application'}
-4. Paste the content (Cmd+V / Ctrl+V)
-5. Save as: ${filename}
-
-## ${formatUpper} Data:
-
-${format === 'xlsx' ?
-    `Excel file exported successfully. ${saveFile && !fileSaveError ?
-      `File has been saved to: ${savedFilePath}\nCompatible with: Excel, Google Sheets, LibreOffice Calc, and other spreadsheet applications` :
-      'To save this Excel file:\n1. Set save_file: true in your export_query parameters\n2. The file will be automatically saved to your Downloads folder\n3. Open with Excel, Google Sheets, or any spreadsheet application'
-    }\n\nTechnical Details:\n- Binary Data: Contains Excel binary data (.xlsx format)\n- High Capacity: Supports up to 1 million rows (vs. 2,000 row limit of standard queries)\n- Native Format: Preserves data types and formatting for spreadsheet applications` :
-    '```' + format + '\n'
-}`;
   }
 
   /**
@@ -826,11 +740,10 @@ ${format === 'xlsx' ?
 
     this.server.setRequestHandler(CallToolRequestSchema, async (request) => {
       const toolName = request.params?.name || 'unknown';
-      const requestId = this.generateRequestId();
+      const requestId = generateRequestId();
 
       this.logInfo(`Processing tool execution request: ${toolName}`, {
         requestId,
-        toolName,
         arguments: request.params?.arguments
       });
 
@@ -838,781 +751,28 @@ ${format === 'xlsx' ?
 
       try {
         switch (request.params?.name) {
-        case 'list_dashboards': {
-          this.logDebug('Fetching all dashboards from Metabase');
-          const response = await this.request<any[]>('/api/dashboard');
-          this.logInfo(`Successfully retrieved ${response.length} dashboards`);
-
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify(response, null, 2)
-            }]
-          };
-        }
-
-        case 'list_cards': {
-          this.logDebug('Fetching all cards/questions from Metabase');
-          const response = await this.getAllCards();
-          this.logInfo(`Successfully retrieved ${response.length} cards/questions`);
-
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify(response, null, 2)
-            }]
-          };
-        }
-
-        case 'list_databases': {
-          this.logDebug('Fetching all databases from Metabase');
-          const response = await this.request<any[]>('/api/database');
-          this.logInfo(`Successfully retrieved ${response.length} databases`);
-
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify(response, null, 2)
-            }]
-          };
-        }
-
-        case 'get_card_sql': {
-          const cardId = request.params?.arguments?.card_id as number;
-          if (!cardId) {
-            this.logWarn('Missing card_id parameter in get_card_sql request', { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'Card ID parameter is required'
-            );
-          }
-
-          this.logDebug(`Fetching SQL details for card with ID: ${cardId}`);
-          const card = await this.request<any>(`/api/card/${cardId}`);
-
-          // Extract relevant information for query execution
-          const result: any = {
-            card_id: cardId,
-            card_name: card.name,
-            database_id: card.database_id,
-            sql_query: card.dataset_query?.native?.query || null,
-            template_tags: card.dataset_query?.native?.template_tags || {},
-            query_type: card.dataset_query?.type || 'unknown',
-            description: card.description || null,
-            collection_id: card.collection_id,
-            created_at: card.created_at,
-            updated_at: card.updated_at
-          };
-
-          // Add guidance for AI agents
-          if (!result.sql_query) {
-            result.message = 'This card does not contain a native SQL query. It may be a GUI-based question.';
-          } else {
-            result.message = 'Use the database_id and sql_query with execute_query. You can modify the SQL query to add filters or parameters.';
-          }
-
-          this.logInfo(`Successfully retrieved SQL details for card: ${cardId}`);
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify(result, null, 2)
-            }]
-          };
-        }
-
-        case 'execute_card': {
-          const cardId = request.params?.arguments?.card_id as number;
-          if (!cardId) {
-            this.logWarn('Missing card_id parameter in execute_card request', { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'Card ID parameter is required'
-            );
-          }
-
-          this.logDebug(`Executing card with ID: ${cardId}`);
-          const parameters = request.params?.arguments?.parameters || {};
-
-          // Convert parameters to the format Metabase expects
-          let formattedParameters: any[] = [];
-
-          if (typeof parameters === 'object' && parameters !== null) {
-            if (Array.isArray(parameters)) {
-              // If already an array, use as-is
-              formattedParameters = parameters;
-            } else {
-              // Convert object format to array format
-              formattedParameters = Object.entries(parameters).map(([key, value]) => {
-                // Determine parameter type based on value
-                let paramType = 'text'; // default type used by Metabase
-                if (typeof value === 'number' || (typeof value === 'string' && /^\d+$/.test(value))) {
-                  paramType = 'id'; // Use 'id' for numeric values (like IDs)
-                } else if (typeof value === 'boolean') {
-                  paramType = 'text';
-                }
-
-                return {
-                  type: paramType,
-                  target: ['variable', ['template-tag', key]], // Correct format: ["variable", ["template-tag", "variable_name"]]
-                  value: value
-                };
-              });
-            }
-          }
-
-          const response = await this.request<any>(`/api/card/${cardId}/query`, {
-            method: 'POST',
-            body: JSON.stringify({ parameters: formattedParameters })
-          });
-
-          this.logInfo(`Successfully executed card: ${cardId}`);
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify(response, null, 2)
-            }]
-          };
-        }
-
-        case 'get_dashboard_cards': {
-          const dashboardId = request.params?.arguments?.dashboard_id;
-          if (!dashboardId) {
-            this.logWarn('Missing dashboard_id parameter in get_dashboard_cards request', { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'Dashboard ID parameter is required'
-            );
-          }
-
-          this.logDebug(`Fetching cards for dashboard with ID: ${dashboardId}`);
-          const response = await this.request<any>(`/api/dashboard/${dashboardId}`);
-
-          const cardCount = response.cards?.length || 0;
-          this.logInfo(`Successfully retrieved ${cardCount} cards from dashboard: ${dashboardId}`);
-
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify(response.cards, null, 2)
-            }]
-          };
-        }
-
-        case 'execute_query': {
-          const databaseId = request.params?.arguments?.database_id;
-          const query = request.params?.arguments?.query;
-          const nativeParameters = request.params?.arguments?.native_parameters || [];
-
-          if (!databaseId) {
-            this.logWarn('Missing database_id parameter in execute_query request', { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'Database ID parameter is required'
-            );
-          }
-
-          if (!query) {
-            this.logWarn('Missing query parameter in execute_query request', { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'SQL query parameter is required'
-            );
-          }
-
-          this.logDebug(`Executing SQL query against database ID: ${databaseId}`);
-
-          // Build query request body
-          const queryData = {
-            type: 'native',
-            native: {
-              query: query,
-              template_tags: {}
-            },
-            parameters: nativeParameters,
-            database: databaseId
-          };
-
-          const response = await this.request<any>('/api/dataset', {
-            method: 'POST',
-            body: JSON.stringify(queryData)
-          });
-
-          this.logInfo(`Successfully executed SQL query against database: ${databaseId}`);
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify(response, null, 2)
-            }]
-          };
-        }
-
-        case 'search_cards': {
-          const searchQuery = request.params?.arguments?.query as string;
-          const searchType = (request.params?.arguments?.search_type as string) || 'auto';
-          const fuzzyThreshold = (request.params?.arguments?.fuzzy_threshold as number) || 0.4;
-          const maxResults = (request.params?.arguments?.max_results as number) || 50;
-
-          if (!searchQuery) {
-            this.logWarn('Missing query parameter in search_cards request', { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'Search query parameter is required'
-            );
-          }
-
-          this.logDebug(`Searching for cards with query: "${searchQuery}" (type: ${searchType}, fuzzy_threshold: ${fuzzyThreshold})`);
-
-          // Fetch all cards using cached method
-          const fetchStartTime = Date.now();
-          const allCards = await this.getAllCards();
-          const fetchTime = Date.now() - fetchStartTime;
-
-          let results: any[] = [];
-          const searchStartTime = Date.now();
-
-          // Determine search type
-          const isNumeric = /^\d+$/.test(searchQuery.trim());
-          let effectiveSearchType = searchType;
-
-          if (searchType === 'auto') {
-            // Auto-detect search type based on query content
-            if (isNumeric) {
-              effectiveSearchType = 'id';
-            } else {
-              // Default to intelligent hybrid search
-              effectiveSearchType = 'auto';
-            }
-          }
-
-          if (effectiveSearchType === 'id') {
-            const targetId = parseInt(searchQuery.trim(), 10);
-            results = allCards.filter(card => card.id === targetId);
-            this.logInfo(`Found ${results.length} cards matching ID: ${targetId}`);
-          } else if (effectiveSearchType === 'exact') {
-            // Exact phrase search
-            const exactResults = this.performExactSearch(
-              allCards,
-              searchQuery,
-              (card) => ({
-                name: card.name,
-                description: card.description,
-                sql: card.dataset_query?.native?.query
-              }),
-              maxResults
-            );
-            results = exactResults;
-            this.logInfo(`Found ${results.length} cards with exact phrase matching: "${searchQuery}"`);
-          } else {
-            // Auto: Intelligent hybrid search (exact + substring + fuzzy)
-            const hybridResults = this.performHybridSearch(
-              allCards,
-              searchQuery,
-              (card) => ({
-                name: card.name,
-                description: card.description,
-                sql: card.dataset_query?.native?.query
-              }),
-              fuzzyThreshold,
-              maxResults
-            );
-            results = hybridResults;
-            this.logInfo(`Found ${results.length} cards using intelligent hybrid search (exact + substring + fuzzy)`);
-          }
-
-          const searchTime = Date.now() - searchStartTime;
-
-          // Enhance results with SQL preview and search matching info
-          const enhancedResults = results.map(card => {
-            const baseCard = {
-              ...card,
-              has_sql: !!(card.dataset_query?.native?.query),
-              sql_preview: card.dataset_query?.native?.query ?
-                card.dataset_query.native.query.substring(0, 200) + (card.dataset_query.native.query.length > 200 ? '...' : '') :
-                null,
-              recommended_action: card.dataset_query?.native?.query ?
-                `Use get_card_sql(${card.id}) then execute_query() for reliable execution` :
-                'This card uses GUI query builder - execute_card may be needed'
-            };
-
-            // Add search matching info based on search type
-            if (effectiveSearchType === 'auto' && 'search_score' in card) {
-              return {
-                ...baseCard,
-                search_score: card.search_score,
-                match_type: card.match_type,
-                matched_field: card.matched_field,
-                match_quality: card.search_score > 0.9 ? 'excellent' :
-                  card.search_score > 0.8 ? 'very good' :
-                    card.search_score > 0.6 ? 'good' : 'moderate'
-              };
-            } else if (effectiveSearchType === 'exact' && 'matched_field' in card) {
-              return {
-                ...baseCard,
-                matched_field: card.matched_field,
-                match_type: 'exact'
-              };
-            }
-
-            return baseCard;
-          });
-
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify({
-                search_query: searchQuery,
-                search_type: effectiveSearchType,
-                fuzzy_threshold: effectiveSearchType === 'fuzzy' ? fuzzyThreshold : undefined,
-                total_results: results.length,
-                performance_info: {
-                  fetch_time_ms: fetchTime,
-                  search_time_ms: searchTime,
-                  total_cards_searched: allCards.length,
-                  cache_used: fetchTime < 1000, // Assume cache was used if fetch was very fast
-                  search_method_used: effectiveSearchType
-                },
-                recommended_workflow: 'For cards with SQL: 1) Use get_card_sql() to get the SQL, 2) Modify if needed, 3) Use execute_query()',
-                search_info: effectiveSearchType === 'auto' ? {
-                  explanation: 'Intelligent hybrid search combining exact matches, substring matches, and fuzzy matching. Results ranked by relevance score.',
-                  fuzzy_threshold_used: fuzzyThreshold,
-                  scoring: 'excellent (>0.9), very good (0.8-0.9), good (0.6-0.8), moderate (0.4-0.6)',
-                  fields_searched: ['name', 'description', 'sql_content'],
-                  match_types: ['exact', 'substring', 'fuzzy']
-                } : effectiveSearchType === 'exact' ? {
-                  explanation: 'Exact phrase matching across all searchable fields.',
-                  fields_searched: ['name', 'description', 'sql_content']
-                } : undefined,
-                results: enhancedResults
-              }, null, 2)
-            }]
-          };
-        }
-
-        case 'search_dashboards': {
-          const searchQuery = request.params?.arguments?.query as string;
-          const searchType = (request.params?.arguments?.search_type as string) || 'auto';
-          const fuzzyThreshold = (request.params?.arguments?.fuzzy_threshold as number) || 0.4;
-          const maxResults = (request.params?.arguments?.max_results as number) || 50;
-
-          if (!searchQuery) {
-            this.logWarn('Missing query parameter in search_dashboards request', { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'Search query parameter is required'
-            );
-          }
-
-          this.logDebug(`Searching for dashboards with query: "${searchQuery}" (type: ${searchType}, fuzzy_threshold: ${fuzzyThreshold})`);
-
-          // Fetch all dashboards first
-          const fetchStartTime = Date.now();
-          const allDashboards = await this.request<any[]>('/api/dashboard');
-          const fetchTime = Date.now() - fetchStartTime;
-
-          let results: any[] = [];
-          const searchStartTime = Date.now();
-
-          // Determine search type
-          const isNumeric = /^\d+$/.test(searchQuery.trim());
-          let effectiveSearchType = searchType;
-
-          if (searchType === 'auto') {
-            // Auto-detect search type based on query content
-            if (isNumeric) {
-              effectiveSearchType = 'id';
-            } else {
-              // Default to intelligent hybrid search
-              effectiveSearchType = 'auto';
-            }
-          }
-
-          if (effectiveSearchType === 'id') {
-            const targetId = parseInt(searchQuery.trim(), 10);
-            results = allDashboards.filter(dashboard => dashboard.id === targetId);
-            this.logInfo(`Found ${results.length} dashboards matching ID: ${targetId}`);
-          } else if (effectiveSearchType === 'exact') {
-            // Exact phrase search
-            const exactResults = this.performExactSearch(
-              allDashboards,
-              searchQuery,
-              (dashboard) => ({
-                name: dashboard.name,
-                description: dashboard.description
-              }),
-              maxResults
-            );
-            results = exactResults;
-            this.logInfo(`Found ${results.length} dashboards with exact phrase matching: "${searchQuery}"`);
-          } else {
-            // Auto: Intelligent hybrid search (exact + substring + fuzzy)
-            const hybridResults = this.performHybridSearch(
-              allDashboards,
-              searchQuery,
-              (dashboard) => ({
-                name: dashboard.name,
-                description: dashboard.description
-              }),
-              fuzzyThreshold,
-              maxResults
-            );
-            results = hybridResults;
-            this.logInfo(`Found ${results.length} dashboards using intelligent hybrid search (exact + substring + fuzzy)`);
-          }
-
-          const searchTime = Date.now() - searchStartTime;
-
-          // Enhance results with search matching info
-          const enhancedResults = results.map(dashboard => {
-            const baseDashboard = { ...dashboard };
-
-            // Add search matching info based on search type
-            if (effectiveSearchType === 'auto' && 'search_score' in dashboard) {
-              return {
-                ...baseDashboard,
-                search_score: dashboard.search_score,
-                match_type: dashboard.match_type,
-                matched_field: dashboard.matched_field,
-                match_quality: dashboard.search_score > 0.9 ? 'excellent' :
-                  dashboard.search_score > 0.8 ? 'very good' :
-                    dashboard.search_score > 0.6 ? 'good' : 'moderate'
-              };
-            } else if (effectiveSearchType === 'exact' && 'matched_field' in dashboard) {
-              return {
-                ...baseDashboard,
-                matched_field: dashboard.matched_field,
-                match_type: 'exact'
-              };
-            }
-
-            return baseDashboard;
-          });
-
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify({
-                search_query: searchQuery,
-                search_type: effectiveSearchType,
-                fuzzy_threshold: effectiveSearchType === 'fuzzy' ? fuzzyThreshold : undefined,
-                total_results: results.length,
-                performance_info: {
-                  fetch_time_ms: fetchTime,
-                  search_time_ms: searchTime,
-                  total_dashboards_searched: allDashboards.length,
-                  search_method_used: effectiveSearchType
-                },
-                search_info: effectiveSearchType === 'auto' ? {
-                  explanation: 'Intelligent hybrid search combining exact matches, substring matches, and fuzzy matching. Results ranked by relevance score.',
-                  fuzzy_threshold_used: fuzzyThreshold,
-                  scoring: 'excellent (>0.9), very good (0.8-0.9), good (0.6-0.8), moderate (0.4-0.6)',
-                  fields_searched: ['name', 'description'],
-                  match_types: ['exact', 'substring', 'fuzzy']
-                } : effectiveSearchType === 'exact' ? {
-                  explanation: 'Exact phrase matching across all searchable fields.',
-                  fields_searched: ['name', 'description']
-                } : undefined,
-                results: enhancedResults
-              }, null, 2)
-            }]
-          };
-        }
-
-        case 'export_query': {
-          const databaseId = request.params?.arguments?.database_id as number;
-          const query = request.params?.arguments?.query as string;
-          const format = (request.params?.arguments?.format as string) || 'csv';
-          const nativeParameters = request.params?.arguments?.native_parameters || [];
-          const saveFile = this.toBooleanSafe(request.params?.arguments?.save_file);
-          const customFilename = request.params?.arguments?.filename as string;
-
-          if (!databaseId) {
-            this.logWarn('Missing database_id parameter in export_query request', { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'Database ID parameter is required'
-            );
-          }
-
-          if (!query) {
-            this.logWarn('Missing query parameter in export_query request', { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'SQL query parameter is required'
-            );
-          }
-
-          if (!['csv', 'json', 'xlsx'].includes(format)) {
-            this.logWarn(`Invalid format parameter in export_query request: ${format}`, { requestId });
-            throw new McpError(
-              ErrorCode.InvalidParams,
-              'Format must be one of: csv, json, xlsx'
-            );
-          }
-
-          this.logDebug(`Exporting query in ${format} format from database ID: ${databaseId}`);
-
-          try {
-            // Build query request body according to Metabase export API requirements
-            const queryData = {
-              type: 'native',
-              native: {
-                query: query,
-                template_tags: {}
-              },
-              parameters: nativeParameters,
-              database: databaseId
-            };
-
-            // Use the export endpoint which supports larger result sets (up to 1M rows)
-            const exportEndpoint = `/api/dataset/${format}`;
-
-            // Build the request body with required parameters as per API documentation
-            const requestBody = {
-              query: queryData,
-              format_rows: false,
-              pivot_results: false,
-              visualization_settings: {}
-            };
-
-            // For export endpoints, we need to handle different response types
-            const url = new URL(exportEndpoint, this.baseUrl);
-            const headers = { ...this.headers };
-
-            // Add appropriate authentication headers
-            if (this.authMethod === AuthMethod.API_KEY && this.apiKey) {
-              headers['X-API-KEY'] = this.apiKey;
-            } else if (this.authMethod === AuthMethod.SESSION && this.sessionToken) {
-              headers['X-Metabase-Session'] = this.sessionToken;
-            }
-
-            const response = await fetch(url.toString(), {
-              method: 'POST',
-              headers,
-              body: JSON.stringify(requestBody)
-            });
-
-            if (!response.ok) {
-              const errorData = await response.json().catch(() => ({}));
-              const errorMessage = `Export API request failed with status ${response.status}: ${response.statusText}`;
-              this.logWarn(errorMessage, errorData);
-              throw {
-                status: response.status,
-                message: response.statusText,
-                data: errorData
-              };
-            }
-
-            // Handle different response types based on format
-            let responseData;
-            if (format === 'json') {
-              responseData = await response.json();
-            } else if (format === 'csv') {
-              // For CSV, get as text
-              responseData = await response.text();
-            } else if (format === 'xlsx') {
-              // For XLSX, get as buffer for binary data
-              responseData = await response.arrayBuffer();
-            } else {
-              // Fallback to text
-              responseData = await response.text();
-            }
-
-            this.logInfo(`Successfully exported query in ${format} format from database: ${databaseId}`);
-
-            if (format === 'json') {
-              // Count rows for user info (JSON format has different structure)
-              const rowCount = responseData?.data?.rows?.length || 0;
-              const timestamp = new Date().toISOString().slice(0, 19).replace(/:/g, '-');
-              const sanitizedCustomFilename = this.sanitizeFilename(customFilename);
-              const baseFilename = sanitizedCustomFilename || `metabase_export_${timestamp}`;
-              const filename = `${baseFilename}.json`;
-
-              let fileSaveError: string | undefined;
-              let savedFilePath = '';
-
-              // Save file if requested
-              if (saveFile) {
-                try {
-                  // Get Downloads directory path
-                  const downloadsPath = path.join(os.homedir(), 'Downloads');
-                  savedFilePath = path.join(downloadsPath, filename);
-
-                  // Ensure Downloads directory exists
-                  if (!fs.existsSync(downloadsPath)) {
-                    fs.mkdirSync(downloadsPath, { recursive: true });
-                  }
-
-                  // Write the JSON file
-                  fs.writeFileSync(savedFilePath, JSON.stringify(responseData, null, 2), 'utf8');
-                } catch (error) {
-                  fileSaveError = error instanceof Error ? error.message : 'Unknown error';
-                }
-              }
-
-              const baseMessage = this.generateExportMessage(
-                format,
-                query,
-                databaseId,
-                rowCount,
-                '',
-                saveFile,
-                savedFilePath,
-                filename,
-                fileSaveError
-              );
-
-              return {
-                content: [{
-                  type: 'text',
-                  text: baseMessage + JSON.stringify(responseData, null, 2) + '\n```\n\nSuccess: Exported ' + rowCount.toLocaleString() + ' rows using Metabase\'s high-capacity export endpoint.\nAdvantage: This method supports up to 1 million rows vs. the 2,000 row limit of standard queries.'
-                }]
-              };
-            } else if (format === 'csv') {
-              // Count rows for user info
-              const rows = responseData.split('\n').filter((row: string) => row.trim());
-              const rowCount = Math.max(0, rows.length - 1); // Subtract header row
-              const timestamp = new Date().toISOString().slice(0, 19).replace(/:/g, '-');
-              const sanitizedCustomFilename = this.sanitizeFilename(customFilename);
-              const baseFilename = sanitizedCustomFilename || `metabase_export_${timestamp}`;
-              const filename = `${baseFilename}.csv`;
-
-              let fileSaveError: string | undefined;
-              let savedFilePath = '';
-
-              // Save file if requested
-              if (saveFile) {
-                try {
-                  // Get Downloads directory path
-                  const downloadsPath = path.join(os.homedir(), 'Downloads');
-                  savedFilePath = path.join(downloadsPath, filename);
-
-                  // Ensure Downloads directory exists
-                  if (!fs.existsSync(downloadsPath)) {
-                    fs.mkdirSync(downloadsPath, { recursive: true });
-                  }
-
-                  // Write the CSV file
-                  fs.writeFileSync(savedFilePath, responseData, 'utf8');
-                } catch (error) {
-                  fileSaveError = error instanceof Error ? error.message : 'Unknown error';
-                }
-              }
-
-              const baseMessage = this.generateExportMessage(
-                format,
-                query,
-                databaseId,
-                rowCount,
-                '',
-                saveFile,
-                savedFilePath,
-                filename,
-                fileSaveError
-              );
-
-              return {
-                content: [{
-                  type: 'text',
-                  text: baseMessage + responseData + '\n```\n\nSuccess: Exported ' + rowCount.toLocaleString() + ' rows using Metabase\'s high-capacity export endpoint.\nAdvantage: This method supports up to 1 million rows vs. the 2,000 row limit of standard queries.'
-                }]
-              };
-            } else {
-              // For XLSX format, handle binary data
-              const timestamp = new Date().toISOString().slice(0, 19).replace(/:/g, '-');
-              const sanitizedCustomFilename = this.sanitizeFilename(customFilename);
-              const baseFilename = sanitizedCustomFilename || `metabase_export_${timestamp}`;
-              const filename = `${baseFilename}.xlsx`;
-
-              let fileSaveError: string | undefined;
-              let savedFilePath = '';
-              let fileSize = 'Unknown';
-
-              // Save file if requested
-              if (saveFile) {
-                try {
-                  // Get Downloads directory path
-                  const downloadsPath = path.join(os.homedir(), 'Downloads');
-                  savedFilePath = path.join(downloadsPath, filename);
-
-                  // Ensure Downloads directory exists
-                  if (!fs.existsSync(downloadsPath)) {
-                    fs.mkdirSync(downloadsPath, { recursive: true });
-                  }
-
-                  // For XLSX, we need to handle binary data properly
-                  // The response should be an ArrayBuffer
-                  if (responseData instanceof ArrayBuffer) {
-                    // Convert ArrayBuffer to Buffer for Node.js file writing
-                    const buffer = Buffer.from(responseData);
-                    fs.writeFileSync(savedFilePath, buffer);
-                    fileSize = buffer.length.toLocaleString();
-                  } else {
-                    // Fallback for other data types
-                    fs.writeFileSync(savedFilePath, responseData);
-                    fileSize = 'Unknown size';
-                  }
-                } catch (error) {
-                  fileSaveError = error instanceof Error ? error.message : 'Unknown error';
-                }
-              } else {
-                if (responseData instanceof ArrayBuffer) {
-                  fileSize = responseData.byteLength.toLocaleString();
-                } else {
-                  fileSize = 'Unknown size';
-                }
-              }
-
-              const baseMessage = this.generateExportMessage(
-                format,
-                query,
-                databaseId,
-                0, // XLSX doesn't have easy row counting
-                fileSize,
-                saveFile,
-                savedFilePath,
-                filename,
-                fileSaveError
-              );
-
-              return {
-                content: [{
-                  type: 'text',
-                  text: baseMessage + '\n\nSuccess: Excel file exported using Metabase\'s high-capacity export endpoint.\nAdvantage: This method supports up to 1 million rows vs. the 2,000 row limit of standard queries.'
-                }]
-              };
-            }
-          } catch (error) {
-            const apiError = error as ApiError;
-            const errorMessage = apiError.data?.message || apiError.message || 'Unknown error';
-
-            this.logError(`Failed to export query in ${format} format: ${errorMessage}`, error);
-            return {
-              content: [{
-                type: 'text',
-                text: `Failed to export query in ${format} format: ${errorMessage}\n\nNote: Make sure your query is valid and the database connection is working. The export endpoint supports up to 1 million rows.`
-              }],
-              isError: true
-            };
-          }
-        }
-
-        case 'clear_cache': {
-          this.logDebug('Clearing cards cache');
-          this.clearCardsCache();
-          this.logInfo('Cards cache cleared successfully');
-
-          return {
-            content: [{
-              type: 'text',
-              text: JSON.stringify({
-                message: 'Cards cache cleared successfully',
-                cache_status: 'empty',
-                next_fetch_will_be: 'fresh from API'
-              }, null, 2)
-            }]
-          };
-        }
-
+        case 'list_dashboards':
+          return this._handleListDashboards();
+        case 'list_cards':
+          return this._handleListCards();
+        case 'list_databases':
+          return this._handleListDatabases();
+        case 'get_card_sql':
+          return this._handleGetCardSql(request, requestId);
+        case 'execute_card':
+          return this._handleExecuteCard(request, requestId);
+        case 'get_dashboard_cards':
+          return this._handleGetDashboardCards(request, requestId);
+        case 'execute_query':
+          return this._handleExecuteQuery(request, requestId);
+        case 'search_cards':
+          return this._handleSearchCards(request, requestId);
+        case 'search_dashboards':
+          return this._handleSearchDashboards(request, requestId);
+        case 'export_query':
+          return this._handleExportQuery(request, requestId);
+        case 'clear_cache':
+          return this._handleClearCache();
         default:
           this.logWarn(`Received request for unknown tool: ${request.params?.name}`, { requestId });
           return {
@@ -1641,231 +801,779 @@ ${format === 'xlsx' ?
     });
   }
 
-  /**
-   * Calculate Levenshtein distance between two strings for fuzzy matching
-   */
-  private calculateLevenshteinDistance(str1: string, str2: string): number {
-    const matrix: number[][] = [];
-    const len1 = str1.length;
-    const len2 = str2.length;
+  private async _handleListDashboards() {
+    this.logDebug('Fetching all dashboards from Metabase');
+    const response = await this.request<any[]>('/api/dashboard');
+    this.logInfo(`Successfully retrieved ${response.length} dashboards`);
 
-    // Initialize matrix
-    for (let i = 0; i <= len1; i++) {
-      matrix[i] = [i];
-    }
-    for (let j = 0; j <= len2; j++) {
-      matrix[0][j] = j;
-    }
-
-    // Fill matrix
-    for (let i = 1; i <= len1; i++) {
-      for (let j = 1; j <= len2; j++) {
-        if (str1[i - 1] === str2[j - 1]) {
-          matrix[i][j] = matrix[i - 1][j - 1];
-        } else {
-          matrix[i][j] = Math.min(
-            matrix[i - 1][j] + 1,     // deletion
-            matrix[i][j - 1] + 1,     // insertion
-            matrix[i - 1][j - 1] + 1  // substitution
-          );
-        }
-      }
-    }
-
-    return matrix[len1][len2];
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify(response, null, 2)
+      }]
+    };
   }
 
-  /**
-   * Calculate fuzzy match score (0-1, where 1 is perfect match)
-   */
-  private calculateFuzzyScore(query: string, target: string): number {
-    if (!query || !target) return 0;
+  private async _handleListCards() {
+    this.logDebug('Fetching all cards/questions from Metabase');
+    const response = await this.getAllCards();
+    this.logInfo(`Successfully retrieved ${response.length} cards/questions`);
 
-    const queryLower = query.toLowerCase().trim();
-    const targetLower = target.toLowerCase().trim();
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify(response, null, 2)
+      }]
+    };
+  }
 
-    // Exact match gets highest score
-    if (queryLower === targetLower) return 1.0;
+  private async _handleListDatabases() {
+    this.logDebug('Fetching all databases from Metabase');
+    const response = await this.request<any[]>('/api/database');
+    this.logInfo(`Successfully retrieved ${response.length} databases`);
 
-    // Check for word-based matches (split by spaces and check individual words)
-    const queryWords = queryLower.split(/\s+/);
-    const targetWords = targetLower.split(/\s+/);
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify(response, null, 2)
+      }]
+    };
+  }
 
-    // Direct contains match gets high score
-    if (targetLower.includes(queryLower)) {
-      const ratio = queryLower.length / targetLower.length;
-      return 0.85 + (ratio * 0.1); // 0.85-0.95 range for contains matches
+  private async _handleGetCardSql(request: z.infer<typeof CallToolRequestSchema>, requestId: string) {
+    const cardId = request.params?.arguments?.card_id as number;
+    if (!cardId) {
+      this.logWarn('Missing card_id parameter in get_card_sql request', { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'Card ID parameter is required'
+      );
     }
 
-    // Check if query matches any complete word in target
-    for (const queryWord of queryWords) {
-      if (queryWord.length >= 3) { // Only check meaningful words
-        for (const targetWord of targetWords) {
-          if (targetWord === queryWord) {
-            return 0.8; // High score for exact word match
+    this.logDebug(`Fetching SQL details for card with ID: ${cardId}`);
+    const card = await this.request<any>(`/api/card/${cardId}`);
+
+    // Extract relevant information for query execution
+    const result: any = {
+      card_id: cardId,
+      card_name: card.name,
+      database_id: card.database_id,
+      sql_query: card.dataset_query?.native?.query || null,
+      template_tags: card.dataset_query?.native?.template_tags || {},
+      query_type: card.dataset_query?.type || 'unknown',
+      description: card.description || null,
+      collection_id: card.collection_id,
+      created_at: card.created_at,
+      updated_at: card.updated_at
+    };
+
+    // Add guidance for AI agents
+    if (!result.sql_query) {
+      result.message = 'This card does not contain a native SQL query. It may be a GUI-based question.';
+    } else {
+      result.message = 'Use the database_id and sql_query with execute_query. You can modify the SQL query to add filters or parameters.';
+    }
+
+    this.logInfo(`Successfully retrieved SQL details for card: ${cardId}`);
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify(result, null, 2)
+      }]
+    };
+  }
+
+  private async _handleExecuteCard(request: z.infer<typeof CallToolRequestSchema>, requestId: string) {
+    const cardId = request.params?.arguments?.card_id as number;
+    if (!cardId) {
+      this.logWarn('Missing card_id parameter in execute_card request', { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'Card ID parameter is required'
+      );
+    }
+
+    this.logDebug(`Executing card with ID: ${cardId}`);
+    const parameters = request.params?.arguments?.parameters || {};
+
+    // Convert parameters to the format Metabase expects
+    let formattedParameters: any[] = [];
+
+    if (typeof parameters === 'object' && parameters !== null) {
+      if (Array.isArray(parameters)) {
+        // If already an array, use as-is
+        formattedParameters = parameters;
+      } else {
+        // Convert object format to array format
+        formattedParameters = Object.entries(parameters).map(([key, value]) => {
+          // Determine parameter type based on value
+          let paramType = 'text'; // default type used by Metabase
+          if (typeof value === 'number' || (typeof value === 'string' && /^\d+$/.test(value))) {
+            paramType = 'id'; // Use 'id' for numeric values (like IDs)
+          } else if (typeof value === 'boolean') {
+            paramType = 'text';
           }
-          if (targetWord.includes(queryWord) && queryWord.length >= 4) {
-            return 0.75; // Good score for word contains
-          }
-        }
+
+          return {
+            type: paramType,
+            target: ['variable', ['template-tag', key]], // Correct format: ["variable", ["template-tag", "variable_name"]]
+            value: value
+          };
+        });
       }
     }
 
-    // Try fuzzy matching on individual words for better typo handling
-    let bestWordScore = 0;
-    for (const queryWord of queryWords) {
-      if (queryWord.length >= 3) {
-        for (const targetWord of targetWords) {
-          if (targetWord.length >= 3) {
-            const distance = this.calculateLevenshteinDistance(queryWord, targetWord);
-            const maxLength = Math.max(queryWord.length, targetWord.length);
-            const wordSimilarity = 1 - (distance / maxLength);
+    const response = await this.request<any>(`/api/card/${cardId}/query`, {
+      method: 'POST',
+      body: JSON.stringify({ parameters: formattedParameters })
+    });
 
-            // Boost score for similar-length words
-            const lengthDiff = Math.abs(queryWord.length - targetWord.length);
-            const lengthPenalty = lengthDiff / maxLength * 0.2;
-            const adjustedScore = Math.max(0, wordSimilarity - lengthPenalty);
+    this.logInfo(`Successfully executed card: ${cardId}`);
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify(response, null, 2)
+      }]
+    };
+  }
 
-            if (adjustedScore > bestWordScore) {
-              bestWordScore = adjustedScore;
+  private async _handleGetDashboardCards(request: z.infer<typeof CallToolRequestSchema>, requestId: string) {
+    const dashboardId = request.params?.arguments?.dashboard_id;
+    if (!dashboardId) {
+      this.logWarn('Missing dashboard_id parameter in get_dashboard_cards request', { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'Dashboard ID parameter is required'
+      );
+    }
+
+    this.logDebug(`Fetching cards for dashboard with ID: ${dashboardId}`);
+    const response = await this.request<any>(`/api/dashboard/${dashboardId}`);
+
+    const cardCount = response.cards?.length || 0;
+    this.logInfo(`Successfully retrieved ${cardCount} cards from dashboard: ${dashboardId}`);
+
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify(response.cards, null, 2)
+      }]
+    };
+  }
+
+  private async _handleExecuteQuery(request: z.infer<typeof CallToolRequestSchema>, requestId: string) {
+    const databaseId = request.params?.arguments?.database_id;
+    const query = request.params?.arguments?.query;
+    const nativeParameters = request.params?.arguments?.native_parameters || [];
+
+    if (!databaseId) {
+      this.logWarn('Missing database_id parameter in execute_query request', { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'Database ID parameter is required'
+      );
+    }
+
+    if (!query) {
+      this.logWarn('Missing query parameter in execute_query request', { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'SQL query parameter is required'
+      );
+    }
+
+    this.logDebug(`Executing SQL query against database ID: ${databaseId}`);
+
+    // Build query request body
+    const queryData = {
+      type: 'native',
+      native: {
+        query: query,
+        template_tags: {}
+      },
+      parameters: nativeParameters,
+      database: databaseId
+    };
+
+    const response = await this.request<any>('/api/dataset', {
+      method: 'POST',
+      body: JSON.stringify(queryData)
+    });
+
+    this.logInfo(`Successfully executed SQL query against database: ${databaseId}`);
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify(response, null, 2)
+      }]
+    };
+  }
+
+  private async _handleSearchCards(request: z.infer<typeof CallToolRequestSchema>, requestId: string) {
+    const searchQuery = request.params?.arguments?.query as string;
+    const searchType = (request.params?.arguments?.search_type as string) || 'auto';
+    const fuzzyThreshold = (request.params?.arguments?.fuzzy_threshold as number) || 0.4;
+    const maxResults = (request.params?.arguments?.max_results as number) || 50;
+
+    if (!searchQuery) {
+      this.logWarn('Missing query parameter in search_cards request', { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'Search query parameter is required'
+      );
+    }
+
+    this.logDebug(`Searching for cards with query: "${searchQuery}" (type: ${searchType}, fuzzy_threshold: ${fuzzyThreshold})`);
+
+    // Fetch all cards using cached method
+    const fetchStartTime = Date.now();
+    const allCards = await this.getAllCards();
+    const fetchTime = Date.now() - fetchStartTime;
+
+    let results: any[] = [];
+    const searchStartTime = Date.now();
+
+    // Determine search type
+    const isNumeric = /^\d+$/.test(searchQuery.trim());
+    let effectiveSearchType = searchType;
+
+    if (searchType === 'auto') {
+      // Auto-detect search type based on query content
+      if (isNumeric) {
+        effectiveSearchType = 'id';
+      } else {
+        // Default to intelligent hybrid search
+        effectiveSearchType = 'auto';
+      }
+    }
+
+    if (effectiveSearchType === 'id') {
+      const targetId = parseInt(searchQuery.trim(), 10);
+      results = allCards.filter(card => card.id === targetId);
+      this.logInfo(`Found ${results.length} cards matching ID: ${targetId}`);
+    } else if (effectiveSearchType === 'exact') {
+      // Exact phrase search
+      const exactResults = performExactSearch(
+        allCards,
+        searchQuery,
+        (card) => ({
+          name: card.name,
+          description: card.description,
+          sql: card.dataset_query?.native?.query
+        }),
+        maxResults
+      );
+      results = exactResults;
+      this.logInfo(`Found ${results.length} cards with exact phrase matching: "${searchQuery}"`);
+    } else {
+      // Auto: Intelligent hybrid search (exact + substring + fuzzy)
+      const hybridResults = performHybridSearch(
+        allCards,
+        searchQuery,
+        (card) => ({
+          name: card.name,
+          description: card.description,
+          sql: card.dataset_query?.native?.query
+        }),
+        fuzzyThreshold,
+        maxResults
+      );
+      results = hybridResults;
+      this.logInfo(`Found ${results.length} cards using intelligent hybrid search (exact + substring + fuzzy)`);
+    }
+
+    const searchTime = Date.now() - searchStartTime;
+
+    // Enhance results with SQL preview and search matching info
+    const enhancedResults = results.map(card => {
+      const baseCard = {
+        ...card,
+        has_sql: !!(card.dataset_query?.native?.query),
+        sql_preview: card.dataset_query?.native?.query ?
+          card.dataset_query.native.query.substring(0, 200) + (card.dataset_query.native.query.length > 200 ? '...' : '') :
+          null,
+        recommended_action: card.dataset_query?.native?.query ?
+          `Use get_card_sql(${card.id}) then execute_query() for reliable execution` :
+          'This card uses GUI query builder - execute_card may be needed'
+      };
+
+      // Add search matching info based on search type
+      if (effectiveSearchType === 'auto' && 'search_score' in card) {
+        return {
+          ...baseCard,
+          search_score: card.search_score,
+          match_type: card.match_type,
+          matched_field: card.matched_field,
+          match_quality: card.search_score > 0.9 ? 'excellent' :
+            card.search_score > 0.8 ? 'very good' :
+              card.search_score > 0.6 ? 'good' : 'moderate'
+        };
+      } else if (effectiveSearchType === 'exact' && 'matched_field' in card) {
+        return {
+          ...baseCard,
+          matched_field: card.matched_field,
+          match_type: 'exact'
+        };
+      }
+
+      return baseCard;
+    });
+
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify({
+          search_query: searchQuery,
+          search_type: effectiveSearchType,
+          fuzzy_threshold: effectiveSearchType === 'fuzzy' ? fuzzyThreshold : undefined,
+          total_results: results.length,
+          performance_info: {
+            fetch_time_ms: fetchTime,
+            search_time_ms: searchTime,
+            total_cards_searched: allCards.length,
+            cache_used: fetchTime < 1000, // Assume cache was used if fetch was very fast
+            search_method_used: effectiveSearchType
+          },
+          recommended_workflow: 'For cards with SQL: 1) Use get_card_sql() to get the SQL, 2) Modify if needed, 3) Use execute_query()',
+          search_info: effectiveSearchType === 'auto' ? {
+            explanation: 'Intelligent hybrid search combining exact matches, substring matches, and fuzzy matching. Results ranked by relevance score.',
+            fuzzy_threshold_used: fuzzyThreshold,
+            scoring: 'excellent (>0.9), very good (0.8-0.9), good (0.6-0.8), moderate (0.4-0.6)',
+            fields_searched: ['name', 'description', 'sql_content'],
+            match_types: ['exact', 'substring', 'fuzzy']
+          } : effectiveSearchType === 'exact' ? {
+            explanation: 'Exact phrase matching across all searchable fields.',
+            fields_searched: ['name', 'description', 'sql_content']
+          } : undefined,
+          results: enhancedResults
+        }, null, 2)
+      }]
+    };
+  }
+
+  private async _handleSearchDashboards(request: z.infer<typeof CallToolRequestSchema>, requestId: string) {
+    const searchQuery = request.params?.arguments?.query as string;
+    const searchType = (request.params?.arguments?.search_type as string) || 'auto';
+    const fuzzyThreshold = (request.params?.arguments?.fuzzy_threshold as number) || 0.4;
+    const maxResults = (request.params?.arguments?.max_results as number) || 50;
+
+    if (!searchQuery) {
+      this.logWarn('Missing query parameter in search_dashboards request', { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'Search query parameter is required'
+      );
+    }
+
+    this.logDebug(`Searching for dashboards with query: "${searchQuery}" (type: ${searchType}, fuzzy_threshold: ${fuzzyThreshold})`);
+
+    // Fetch all dashboards first
+    const fetchStartTime = Date.now();
+    const allDashboards = await this.request<any[]>('/api/dashboard');
+    const fetchTime = Date.now() - fetchStartTime;
+
+    let results: any[] = [];
+    const searchStartTime = Date.now();
+
+    // Determine search type
+    const isNumeric = /^\d+$/.test(searchQuery.trim());
+    let effectiveSearchType = searchType;
+
+    if (searchType === 'auto') {
+      // Auto-detect search type based on query content
+      if (isNumeric) {
+        effectiveSearchType = 'id';
+      } else {
+        // Default to intelligent hybrid search
+        effectiveSearchType = 'auto';
+      }
+    }
+
+    if (effectiveSearchType === 'id') {
+      const targetId = parseInt(searchQuery.trim(), 10);
+      results = allDashboards.filter(dashboard => dashboard.id === targetId);
+      this.logInfo(`Found ${results.length} dashboards matching ID: ${targetId}`);
+    } else if (effectiveSearchType === 'exact') {
+      // Exact phrase search
+      const exactResults = performExactSearch(
+        allDashboards,
+        searchQuery,
+        (dashboard) => ({
+          name: dashboard.name,
+          description: dashboard.description
+        }),
+        maxResults
+      );
+      results = exactResults;
+      this.logInfo(`Found ${results.length} dashboards with exact phrase matching: "${searchQuery}"`);
+    } else {
+      // Auto: Intelligent hybrid search (exact + substring + fuzzy)
+      const hybridResults = performHybridSearch(
+        allDashboards,
+        searchQuery,
+        (dashboard) => ({
+          name: dashboard.name,
+          description: dashboard.description
+        }),
+        fuzzyThreshold,
+        maxResults
+      );
+      results = hybridResults;
+      this.logInfo(`Found ${results.length} dashboards using intelligent hybrid search (exact + substring + fuzzy)`);
+    }
+
+    const searchTime = Date.now() - searchStartTime;
+
+    // Enhance results with search matching info
+    const enhancedResults = results.map(dashboard => {
+      const baseDashboard = { ...dashboard };
+
+      // Add search matching info based on search type
+      if (effectiveSearchType === 'auto' && 'search_score' in dashboard) {
+        return {
+          ...baseDashboard,
+          search_score: dashboard.search_score,
+          match_type: dashboard.match_type,
+          matched_field: dashboard.matched_field,
+          match_quality: dashboard.search_score > 0.9 ? 'excellent' :
+            dashboard.search_score > 0.8 ? 'very good' :
+              dashboard.search_score > 0.6 ? 'good' : 'moderate'
+        };
+      } else if (effectiveSearchType === 'exact' && 'matched_field' in dashboard) {
+        return {
+          ...baseDashboard,
+          matched_field: dashboard.matched_field,
+          match_type: 'exact'
+        };
+      }
+
+      return baseDashboard;
+    });
+
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify({
+          search_query: searchQuery,
+          search_type: effectiveSearchType,
+          fuzzy_threshold: effectiveSearchType === 'fuzzy' ? fuzzyThreshold : undefined,
+          total_results: results.length,
+          performance_info: {
+            fetch_time_ms: fetchTime,
+            search_time_ms: searchTime,
+            total_dashboards_searched: allDashboards.length,
+            search_method_used: effectiveSearchType
+          },
+          search_info: effectiveSearchType === 'auto' ? {
+            explanation: 'Intelligent hybrid search combining exact matches, substring matches, and fuzzy matching. Results ranked by relevance score.',
+            fuzzy_threshold_used: fuzzyThreshold,
+            scoring: 'excellent (>0.9), very good (0.8-0.9), good (0.6-0.8), moderate (0.4-0.6)',
+            fields_searched: ['name', 'description'],
+            match_types: ['exact', 'substring', 'fuzzy']
+          } : effectiveSearchType === 'exact' ? {
+            explanation: 'Exact phrase matching across all searchable fields.',
+            fields_searched: ['name', 'description']
+          } : undefined,
+          results: enhancedResults
+        }, null, 2)
+      }]
+    };
+  }
+
+  private async _handleExportQuery(request: z.infer<typeof CallToolRequestSchema>, requestId: string) {
+    const databaseId = request.params?.arguments?.database_id as number;
+    const query = request.params?.arguments?.query as string;
+    const format = (request.params?.arguments?.format as string) || 'csv';
+    const nativeParameters = request.params?.arguments?.native_parameters || [];
+    const saveFile = toBooleanSafe(request.params?.arguments?.save_file);
+    const customFilename = request.params?.arguments?.filename as string;
+
+    if (!databaseId) {
+      this.logWarn('Missing database_id parameter in export_query request', { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'Database ID parameter is required'
+      );
+    }
+
+    if (!query) {
+      this.logWarn('Missing query parameter in export_query request', { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'SQL query parameter is required'
+      );
+    }
+
+    if (!['csv', 'json', 'xlsx'].includes(format)) {
+      this.logWarn(`Invalid format parameter in export_query request: ${format}`, { requestId });
+      throw new McpError(
+        ErrorCode.InvalidParams,
+        'Format must be one of: csv, json, xlsx'
+      );
+    }
+
+    this.logDebug(`Exporting query in ${format} format from database ID: ${databaseId}`);
+
+    try {
+      // Build query request body according to Metabase export API requirements
+      const queryData = {
+        type: 'native',
+        native: {
+          query: query,
+          template_tags: {}
+        },
+        parameters: nativeParameters,
+        database: databaseId
+      };
+
+      // Use the export endpoint which supports larger result sets (up to 1M rows)
+      const exportEndpoint = `/api/dataset/${format}`;
+
+      // Build the request body with required parameters as per API documentation
+      const requestBody = {
+        query: queryData,
+        format_rows: false,
+        pivot_results: false,
+        visualization_settings: {}
+      };
+
+      // For export endpoints, we need to handle different response types
+      const url = new URL(exportEndpoint, this.baseUrl);
+      const headers = { ...this.headers };
+
+      // Add appropriate authentication headers
+      if (this.authMethod === AuthMethod.API_KEY && this.apiKey) {
+        headers['X-API-KEY'] = this.apiKey;
+      } else if (this.authMethod === AuthMethod.SESSION && this.sessionToken) {
+        headers['X-Metabase-Session'] = this.sessionToken;
+      }
+
+      const response = await fetch(url.toString(), {
+        method: 'POST',
+        headers,
+        body: JSON.stringify(requestBody)
+      });
+
+      if (!response.ok) {
+        const errorData = await response.json().catch(() => ({}));
+        const errorMessage = `Export API request failed with status ${response.status}: ${response.statusText}`;
+        this.logWarn(errorMessage, errorData);
+        throw {
+          status: response.status,
+          message: response.statusText,
+          data: errorData
+        };
+      }
+
+      // Handle different response types based on format
+      let responseData;
+      if (format === 'json') {
+        responseData = await response.json();
+      } else if (format === 'csv') {
+        // For CSV, get as text
+        responseData = await response.text();
+      } else if (format === 'xlsx') {
+        // For XLSX, get as buffer for binary data
+        responseData = await response.arrayBuffer();
+      } else {
+        // Fallback to text
+        responseData = await response.text();
+      }
+
+      this.logInfo(`Successfully exported query in ${format} format from database: ${databaseId}`);
+
+      if (format === 'json') {
+        // Count rows for user info (JSON format has different structure)
+        const rowCount = responseData?.data?.rows?.length || 0;
+        const timestamp = new Date().toISOString().slice(0, 19).replace(/:/g, '-');
+        const sanitizedCustomFilename = sanitizeFilename(customFilename);
+        const baseFilename = sanitizedCustomFilename || `metabase_export_${timestamp}`;
+        const filename = `${baseFilename}.json`;
+
+        let fileSaveError: string | undefined;
+        let savedFilePath = '';
+
+        // Save file if requested
+        if (saveFile) {
+          try {
+            // Get Downloads directory path
+            const downloadsPath = path.join(os.homedir(), 'Downloads');
+            savedFilePath = path.join(downloadsPath, filename);
+
+            // Ensure Downloads directory exists
+            if (!fs.existsSync(downloadsPath)) {
+              fs.mkdirSync(downloadsPath, { recursive: true });
             }
-          }
-        }
-      }
-    }
 
-    // If we found a good word match, use it
-    if (bestWordScore >= 0.6) {
-      return Math.min(bestWordScore + 0.1, 0.8); // Cap at 0.8 for fuzzy word matches
-    }
-
-    // Calculate Levenshtein distance for full string fuzzy matching
-    const distance = this.calculateLevenshteinDistance(queryLower, targetLower);
-    const maxLength = Math.max(queryLower.length, targetLower.length);
-
-    // Convert distance to similarity score (0-1)
-    const similarity = 1 - (distance / maxLength);
-
-    // For shorter queries, be more lenient
-    const lengthBonus = queryLower.length <= 5 ? 0.1 : 0;
-    const finalScore = similarity + lengthBonus;
-
-    // Apply threshold - only return scores above 0.4 for fuzzy matching
-    return finalScore >= 0.4 ? Math.min(finalScore, 1.0) : 0;
-  }
-
-
-
-  /**
-   * Perform intelligent hybrid search combining exact, substring, and fuzzy matching
-   */
-  private performHybridSearch<T>(
-    items: T[],
-    query: string,
-    getSearchFields: (item: T) => { name?: string; description?: string; sql?: string },
-    fuzzyThreshold: number = 0.4,
-    maxResults: number = 50
-  ): Array<T & { search_score: number; match_type: string; matched_field: string }> {
-    const results: Array<T & { search_score: number; match_type: string; matched_field: string }> = [];
-    const queryLower = query.toLowerCase().trim();
-
-    for (const item of items) {
-      const fields = getSearchFields(item);
-      let bestScore = 0;
-      let bestMatchType = '';
-      let bestField = '';
-
-      // Check each field (name, description, SQL)
-      for (const [fieldName, fieldValue] of Object.entries(fields)) {
-        if (!fieldValue) continue;
-
-        const fieldLower = fieldValue.toLowerCase();
-        let score = 0;
-        let matchType = '';
-
-        // 1. Exact match (highest priority)
-        if (fieldLower === queryLower) {
-          score = 1.0;
-          matchType = 'exact';
-        }
-        // 2. Substring match (high priority)
-        else if (fieldLower.includes(queryLower)) {
-          // Score based on how much of the field the query represents
-          const ratio = queryLower.length / fieldLower.length;
-          score = 0.85 + (ratio * 0.1); // 0.85-0.95 range
-          matchType = 'substring';
-        }
-        // 3. Fuzzy match (lower priority)
-        else {
-          const fuzzyScore = this.calculateFuzzyScore(query, fieldValue);
-          if (fuzzyScore >= fuzzyThreshold) {
-            score = fuzzyScore;
-            matchType = 'fuzzy';
+            // Write the JSON file
+            fs.writeFileSync(savedFilePath, JSON.stringify(responseData, null, 2), 'utf8');
+          } catch (error) {
+            fileSaveError = error instanceof Error ? error.message : 'Unknown error';
           }
         }
 
-        // Keep the best match for this item
-        if (score > bestScore) {
-          bestScore = score;
-          bestMatchType = matchType;
-          bestField = `${fieldName}: ${fieldValue.substring(0, 50)}${fieldValue.length > 50 ? '...' : ''}`;
+        const baseMessage = generateExportMessage(
+          format,
+          query,
+          databaseId,
+          rowCount,
+          '',
+          saveFile,
+          savedFilePath,
+          filename,
+          fileSaveError
+        );
+
+        return {
+          content: [{
+            type: 'text',
+            text: baseMessage + JSON.stringify(responseData, null, 2) + '\n```\n\nSuccess: Exported ' + rowCount.toLocaleString() + ' rows using Metabase\'s high-capacity export endpoint.\nAdvantage: This method supports up to 1 million rows vs. the 2,000 row limit of standard queries.'
+          }]
+        };
+      } else if (format === 'csv') {
+        // Count rows for user info
+        const rows = responseData.split('\n').filter((row: string) => row.trim());
+        const rowCount = Math.max(0, rows.length - 1); // Subtract header row
+        const timestamp = new Date().toISOString().slice(0, 19).replace(/:/g, '-');
+        const sanitizedCustomFilename = sanitizeFilename(customFilename);
+        const baseFilename = sanitizedCustomFilename || `metabase_export_${timestamp}`;
+        const filename = `${baseFilename}.csv`;
+
+        let fileSaveError: string | undefined;
+        let savedFilePath = '';
+
+        // Save file if requested
+        if (saveFile) {
+          try {
+            // Get Downloads directory path
+            const downloadsPath = path.join(os.homedir(), 'Downloads');
+            savedFilePath = path.join(downloadsPath, filename);
+
+            // Ensure Downloads directory exists
+            if (!fs.existsSync(downloadsPath)) {
+              fs.mkdirSync(downloadsPath, { recursive: true });
+            }
+
+            // Write the CSV file
+            fs.writeFileSync(savedFilePath, responseData, 'utf8');
+          } catch (error) {
+            fileSaveError = error instanceof Error ? error.message : 'Unknown error';
+          }
         }
-      }
 
-      // Add item if it has any match
-      if (bestScore > 0) {
-        results.push({
-          ...item,
-          search_score: bestScore,
-          match_type: bestMatchType,
-          matched_field: bestField
-        });
+        const baseMessage = generateExportMessage(
+          format,
+          query,
+          databaseId,
+          rowCount,
+          '',
+          saveFile,
+          savedFilePath,
+          filename,
+          fileSaveError
+        );
+
+        return {
+          content: [{
+            type: 'text',
+            text: baseMessage + responseData + '\n```\n\nSuccess: Exported ' + rowCount.toLocaleString() + ' rows using Metabase\'s high-capacity export endpoint.\nAdvantage: This method supports up to 1 million rows vs. the 2,000 row limit of standard queries.'
+          }]
+        };
+      } else {
+        // For XLSX format, handle binary data
+        const timestamp = new Date().toISOString().slice(0, 19).replace(/:/g, '-');
+        const sanitizedCustomFilename = sanitizeFilename(customFilename);
+        const baseFilename = sanitizedCustomFilename || `metabase_export_${timestamp}`;
+        const filename = `${baseFilename}.xlsx`;
+
+        let fileSaveError: string | undefined;
+        let savedFilePath = '';
+        let fileSize = 'Unknown';
+
+        // Save file if requested
+        if (saveFile) {
+          try {
+            // Get Downloads directory path
+            const downloadsPath = path.join(os.homedir(), 'Downloads');
+            savedFilePath = path.join(downloadsPath, filename);
+
+            // Ensure Downloads directory exists
+            if (!fs.existsSync(downloadsPath)) {
+              fs.mkdirSync(downloadsPath, { recursive: true });
+            }
+
+            // For XLSX, we need to handle binary data properly
+            // The response should be an ArrayBuffer
+            if (responseData instanceof ArrayBuffer) {
+              // Convert ArrayBuffer to Buffer for Node.js file writing
+              const buffer = Buffer.from(responseData);
+              fs.writeFileSync(savedFilePath, buffer);
+              fileSize = buffer.length.toLocaleString();
+            } else {
+              // Fallback for other data types
+              fs.writeFileSync(savedFilePath, responseData);
+              fileSize = 'Unknown size';
+            }
+          } catch (error) {
+            fileSaveError = error instanceof Error ? error.message : 'Unknown error';
+          }
+        } else {
+          if (responseData instanceof ArrayBuffer) {
+            fileSize = responseData.byteLength.toLocaleString();
+          } else {
+            fileSize = 'Unknown size';
+          }
+        }
+
+        const baseMessage = generateExportMessage(
+          format,
+          query,
+          databaseId,
+          0, // XLSX doesn't have easy row counting
+          fileSize,
+          saveFile,
+          savedFilePath,
+          filename,
+          fileSaveError
+        );
+
+        return {
+          content: [{
+            type: 'text',
+            text: baseMessage + '\n\nSuccess: Excel file exported using Metabase\'s high-capacity export endpoint.\nAdvantage: This method supports up to 1 million rows vs. the 2,000 row limit of standard queries.'
+          }]
+        };
       }
+    } catch (error) {
+      const apiError = error as ApiError;
+      const errorMessage = apiError.data?.message || apiError.message || 'Unknown error';
+
+      this.logError(`Failed to export query in ${format} format: ${errorMessage}`, error);
+      return {
+        content: [{
+          type: 'text',
+          text: `Failed to export query in ${format} format: ${errorMessage}\n\nNote: Make sure your query is valid and the database connection is working. The export endpoint supports up to 1 million rows.`
+        }],
+        isError: true
+      };
     }
-
-    // Sort by score (descending) and limit results
-    return results
-      .sort((a, b) => b.search_score - a.search_score)
-      .slice(0, maxResults);
   }
 
-  /**
-   * Perform exact phrase search
-   */
-  private performExactSearch<T>(
-    items: T[],
-    query: string,
-    getSearchFields: (item: T) => { name?: string; description?: string; sql?: string },
-    maxResults: number = 50
-  ): Array<T & { matched_field: string }> {
-    const results: Array<T & { matched_field: string }> = [];
-    const queryLower = query.toLowerCase().trim();
+  private _handleClearCache() {
+    this.logDebug('Clearing cards cache');
+    this.clearCardsCache();
+    this.logInfo('Cards cache cleared successfully');
 
-    for (const item of items) {
-      const fields = getSearchFields(item);
-      let matchedField = '';
-
-      // Check each field for exact phrase match
-      for (const [fieldName, fieldValue] of Object.entries(fields)) {
-        if (!fieldValue) continue;
-
-        const fieldLower = fieldValue.toLowerCase();
-        if (fieldLower.includes(queryLower)) {
-          matchedField = `${fieldName}: ${fieldValue.substring(0, 50)}${fieldValue.length > 50 ? '...' : ''}`;
-          break; // Take first match
-        }
-      }
-
-      // Add item if exact phrase was found
-      if (matchedField) {
-        results.push({
-          ...item,
-          matched_field: matchedField
-        });
-      }
-    }
-
-    return results.slice(0, maxResults);
+    return {
+      content: [{
+        type: 'text',
+        text: JSON.stringify({
+          message: 'Cards cache cleared successfully',
+          cache_status: 'empty',
+          next_fetch_will_be: 'fresh from API'
+        }, null, 2)
+      }]
+    };
   }
 
   async run() {
